Starts the work of changing display to LCD, serial comms to ASCII and storing controller state to EEPROM

+++ controller/arduino/src/hmi/alarmled.hpp
@@ -1,23 +1,19 @@
 #ifndef ALARMLED_H
 #define ALARMLED_H
 
-#include "Arduino.h"
 #include "../utils/component.hpp"
+#include "Arduino.h"
 
 using namespace std;
 
-class AlarmLED
-    : public Component
-{
-private:
+class AlarmLED : public Component {
+  private:
     int led_pin;
 
-public:
-    AlarmLED(
-        Observable *events,
-        int led_pin);
+  public:
+    AlarmLED(Observable *events, int led_pin);
     void setup();
-    void notify(const char *event, uint16_t payload);
+    void notify(const char *event, int payload);
     void loop();
 };
