+++ controller/arduino/src/pwmfan/pwmfan.hpp
@@ -1,27 +1,21 @@
 #ifndef PWMFAN_H
 #define PWMFAN_H
 
-#include "Arduino.h"
 #include "../utils/timedcomponent.hpp"
+#include "Arduino.h"
 
 using namespace std;
 
-class PWMFan
-    : public TimedComponent
-{
-private:
+class PWMFan : public TimedComponent {
+  private:
     int sense_pin;
     int pwm_pin;
     volatile int fan_pulses;
 
-public:
-    PWMFan(
-        Observable *events,
-        int update_interval,
-        int sense_pin,
-        int pwm_pin);
+  public:
+    PWMFan(Observable *events, int update_interval, int sense_pin, int pwm_pin);
     void setup();
-    void notify(const char *event, uint16_t payload);
+    void notify(const char *event, int payload);
     void update();
     void pickPulse();
 };
