+++ controller/arduino/src/utils/observable.cpp
@@ -7,10 +7,8 @@ using namespace std;
  *
  * @param observer
  */
-void Observable::register_observer(Observer *observer)
-{
-    if (arr_length < 50)
-    {
+void Observable::register_observer(Observer *observer) {
+    if (arr_length < 50) {
         observers[arr_length] = observer;
         arr_length++;
     }
@@ -21,10 +19,8 @@ void Observable::register_observer(Observer *observer)
  *
  * @param payload
  */
-void Observable::notify_observers(const char *event, uint16_t payload)
-{
-    for (int i = 0; i < arr_length; i++)
-    {
+void Observable::notify_observers(const char *event, int payload) {
+    for (int i = 0; i < arr_length; i++) {
         observers[i]->notify(event, payload);
     }
 }
