+++ controller/arduino/src/utils/observable.hpp
@@ -1,18 +1,17 @@
 #ifndef OBSERVABLE_H
 #define OBSERVABLE_H
 
-#include <Arduino.h>
 #include "observer.hpp"
+#include <Arduino.h>
 
-class Observable
-{
-private:
+class Observable {
+  private:
     Observer *observers[50];
     int arr_length;
 
-public:
+  public:
     void register_observer(Observer *observer);
-    void notify_observers(const char *event, uint16_t payload);
+    void notify_observers(const char *event, int payload);
 };
 
 #endif
